[FIX] Int16범위 이상 칼로리입력시 크래시 수정

+++ YangeYammy/YangeYammy/Screen/Profile/ProfileAViewController.swift
@@ -252,7 +252,7 @@ extension ProfileAViewController: UITextFieldDelegate {
         case profileAView.kcal:
             if let kcalText = textField.text,
                 let kcal = Int(kcalText) {
-                profile.kcal = Int16(kcal)
+                profile.kcal = Int64(kcal)
             }
             
         default:
